+++ lib/models/ListingRecord.dart
@@ -1,6 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 
-class Listing {
+class ListingRecord {
   String title;
   GeoPoint geoPoint;
   double pricePerWeek;
@@ -12,22 +12,30 @@ class Listing {
   CollectionReference requestCollection;
   DocumentReference userReference;
 
-  Listing(this.title, this.geoPoint, this.pricePerWeek, this.totalRooms,
-      this.freeRooms, this.genderPreference, this.photoURLs, this.userReference);
+  ListingRecord(
+      this.title,
+      this.geoPoint,
+      this.pricePerWeek,
+      this.totalRooms,
+      this.freeRooms,
+      this.genderPreference,
+      this.photoURLs,
+      this.userReference);
 
-  Listing.fromMap(Map<String, dynamic> map, {this.reference}){
+  ListingRecord.fromMap(Map<String, dynamic> map, {this.reference}) {
     this.title = map["title"];
     this.geoPoint = map["geoPoint"];
     this.pricePerWeek = double.parse(map["pricePerWeek"].toString());
     this.totalRooms = map["totalRooms"];
     this.freeRooms = map["freeRooms"];
-    this.genderPreference = Gender.values.firstWhere((e) => e.toString() == "Gender." + map["genderPreference"]);
+    this.genderPreference = Gender.values
+        .firstWhere((e) => e.toString() == "Gender." + map["genderPreference"]);
     this.photoURLs = map["photoURLs"];
     this.userReference = map["owner"];
     this.requestCollection = map["requests"];
   }
 
-  Listing.fromSnapshot(DocumentSnapshot snapshot)
+  ListingRecord.fromSnapshot(DocumentSnapshot snapshot)
       : this.fromMap(snapshot.data(), reference: snapshot.reference);
 
   Map<String, dynamic> toFirebase() {
@@ -44,9 +52,4 @@ class Listing {
   }
 }
 
-enum Gender {
-  Female,
-  Male,
-  Other,
-  NoPreference
-}
+enum Gender { Female, Male, Other, NoPreference }
