+++ lib/screens/chat/ViewChatPage.dart
@@ -3,11 +3,11 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:uni_roomie/screens/profile/viewOtherProfile.dart';
+import 'package:uni_roomie/screens/profile/ViewOtherProfilePage.dart';
 
-import 'ChatRecord.dart';
-import 'MessageRecord.dart';
-import 'UserRecord.dart';
+import '../../models/ChatRecord.dart';
+import '../../models/MessageRecord.dart';
+import '../../models/UserRecord.dart';
 
 class ViewChatPage extends StatefulWidget {
   final ChatRecord chatRecord;
@@ -74,7 +74,9 @@ class _ViewChatPageState extends State<ViewChatPage> {
                 ),
                 onPressed: () => Navigator.push(
                   context,
-                  MaterialPageRoute(builder: (context) => viewOtherProfilePage(otherUser.reference.id)),
+                  MaterialPageRoute(
+                      builder: (context) =>
+                          ViewOtherProfilePage(otherUser.reference.id)),
                 ),
               ),
               SizedBox(
