+++ lib/screens/chat/ViewChatsPage.dart
@@ -1,14 +1,12 @@
 import 'dart:async';
-import 'dart:io';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:uni_roomie/screens/searchListing/viewListing.dart';
-import 'package:uni_roomie/screens/viewListings/viewListings.dart';
+import 'package:uni_roomie/screens/listing/SearchListingsPage.dart';
 
-import 'ChatRecord.dart';
-import 'UserRecord.dart';
+import '../../models/ChatRecord.dart';
+import '../../models/UserRecord.dart';
 import 'ViewChatPage.dart';
 
 class ViewChatsPage extends StatefulWidget {
@@ -36,53 +34,69 @@ class _ViewChatsPageState extends State<ViewChatsPage> {
         FirebaseFirestore.instance.collection("users").doc(currentUser);
 
     return StreamBuilder<QuerySnapshot>(
-      stream: FirebaseFirestore.instance
-          .collection("chats")
-          .snapshots(),
+      stream: FirebaseFirestore.instance.collection("chats").snapshots(),
       builder: (context, snapshot) {
         if (!snapshot.hasData) return LinearProgressIndicator();
 
-        List<ChatRecord> userRecords = snapshot.data.docs.map((e) => ChatRecord.fromSnapshot(e)).toList();
+        List<ChatRecord> userRecords =
+            snapshot.data.docs.map((e) => ChatRecord.fromSnapshot(e)).toList();
 
-        return _buildList(context, userRecords.where((e) => e.user1 == userReference || e.user2 == userReference).toList());
+        return _buildList(
+            context,
+            userRecords
+                .where(
+                    (e) => e.user1 == userReference || e.user2 == userReference)
+                .toList());
       },
     );
   }
 
-  Widget _buildList(
-      BuildContext context, List<ChatRecord> chatRecords) {
-    if (chatRecords.isEmpty) return Padding(
-      padding: const EdgeInsets.all(16.0),
-      child: Column(children: [Text("No current chats, message a listing owner to get chatting!", textAlign: TextAlign.center, style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20),),              Container(
-        padding: const EdgeInsets.fromLTRB(16.0, 16.0, 16.0, 16.0),
-        child: RaisedButton(
-          shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(18.0)),
-          onPressed: () {Navigator.push(
-            context,
-            MaterialPageRoute(builder: (context) => viewListingPage()),
-          );},
-          color: new Color.fromRGBO(249, 89, 89, 1),
-          child: Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Text(
-                  'View Listings',
-                  style: TextStyle(fontSize: 20.0),
-                ),
-                SizedBox(width: 10),
-                Icon(
-                  Icons.house,
-                  color: Colors.black,
-                ),
-              ],
+  Widget _buildList(BuildContext context, List<ChatRecord> chatRecords) {
+    if (chatRecords.isEmpty)
+      return Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          children: [
+            Text(
+              "No current chats, message a listing owner to get chatting!",
+              textAlign: TextAlign.center,
+              style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20),
             ),
-          ),
+            Container(
+              padding: const EdgeInsets.fromLTRB(16.0, 16.0, 16.0, 16.0),
+              child: RaisedButton(
+                shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(18.0)),
+                onPressed: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                        builder: (context) => SearchListingsPage()),
+                  );
+                },
+                color: new Color.fromRGBO(249, 89, 89, 1),
+                child: Padding(
+                  padding: const EdgeInsets.all(16.0),
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Text(
+                        'View Listings',
+                        style: TextStyle(fontSize: 20.0),
+                      ),
+                      SizedBox(width: 10),
+                      Icon(
+                        Icons.house,
+                        color: Colors.black,
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+            ),
+          ],
         ),
-      ),])
-    );
+      );
     return ListView(
       padding: const EdgeInsets.only(top: 20.0),
       children: chatRecords
@@ -103,7 +117,9 @@ class _ViewChatsPageState extends State<ViewChatsPage> {
         FirebaseFirestore.instance.collection("users").doc(currentUser.uid);
 
     return FutureBuilder(
-        future: getUserInfo(userReference == chatRecord.user1 ? chatRecord.user2 : chatRecord.user1),
+        future: getUserInfo(userReference == chatRecord.user1
+            ? chatRecord.user2
+            : chatRecord.user1),
         builder: (context, AsyncSnapshot<DocumentSnapshot> userSnapshot) {
           if (!userSnapshot.hasData) {
             return LinearProgressIndicator();
