+++ lib/screens/listing/CreateListingPage.dart
@@ -5,27 +5,29 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
+import 'package:google_map_location_picker/google_map_location_picker.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:provider/provider.dart';
-import 'package:uni_roomie/blocs/auth_bloc.dart';
+import 'package:uni_roomie/blocs/AuthBloc.dart';
 import 'package:uni_roomie/customtiles/CustomTile.dart';
-import 'package:uni_roomie/objects/listing.dart';
-import 'package:uni_roomie/screens/login/login.dart';
-import 'package:google_map_location_picker/google_map_location_picker.dart';
+import 'package:uni_roomie/screens/login/LoginPage.dart';
+
+import 'file:///C:/Users/Joe/AndroidStudioProjects/AstonHack2021/lib/models/ListingRecord.dart';
 
 //Divider - > Container(height: 100, child: Divider(color: Colors.black))
 
-class createListingPage extends StatefulWidget {
+class CreateListingPage extends StatefulWidget {
   TextEditingController titleCont = new TextEditingController();
   TextEditingController priceCont = new TextEditingController();
   TextEditingController totalRoomsCont = new TextEditingController();
   TextEditingController freeRoomsCont = new TextEditingController();
   LocationResult result;
+
   @override
-  _createListingPageState createState() => _createListingPageState();
+  _CreateListingPageState createState() => _CreateListingPageState();
 }
 
-class _createListingPageState extends State<createListingPage> {
+class _CreateListingPageState extends State<CreateListingPage> {
   StreamSubscription<User> loginStateSubscription;
   DocumentReference userReference;
 
@@ -98,7 +100,8 @@ class _createListingPageState extends State<createListingPage> {
                     mainAxisAlignment: MainAxisAlignment.spaceBetween,
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
-                      textContainer('Title...', 'Listing Title', widget.titleCont),
+                      textContainer(
+                          'Title...', 'Listing Title', widget.titleCont),
                       SizedBox(height: 50),
                       Container(
                         padding:
@@ -109,7 +112,8 @@ class _createListingPageState extends State<createListingPage> {
                               shape: RoundedRectangleBorder(
                                   borderRadius: BorderRadius.circular(18.0)),
                               onPressed: () async {
-                                widget.result = await showLocationPicker(context,
+                                widget.result = await showLocationPicker(
+                                    context,
                                     "AIzaSyDY6RwTm0Dhr7YMs_jLi6B8fwqhTyCqzJw");
                                 if (widget.result != null) {
                                   setState(() {
@@ -164,7 +168,8 @@ class _createListingPageState extends State<createListingPage> {
                       SizedBox(
                         height: 30,
                       ),
-                      numberContainer('Price...', 'Price Per Week', widget.priceCont),
+                      numberContainer(
+                          'Price...', 'Price Per Week', widget.priceCont),
                       SizedBox(
                         height: 50,
                       ),
@@ -256,7 +261,7 @@ class _createListingPageState extends State<createListingPage> {
                               }
 
                               void insertRow() async {
-                                Listing listing = new Listing(
+                                ListingRecord listing = new ListingRecord(
                                     widget.titleCont.text,
                                     new GeoPoint(widget.result.latLng.latitude,
                                         widget.result.latLng.longitude),
