+++ lib/screens/profile/ProfilePage.dart
@@ -1,20 +1,23 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
+import 'package:image_picker/image_picker.dart';
 import 'package:provider/provider.dart';
-import 'package:uni_roomie/blocs/auth_bloc.dart';
+import 'package:uni_roomie/blocs/AuthBloc.dart';
 import 'package:uni_roomie/customtiles/CustomTile.dart';
-import 'package:uni_roomie/screens/login/login.dart';
-import 'package:uni_roomie/screens/profile/profile.dart';
+import 'package:uni_roomie/screens/login/LoginPage.dart';
+import 'package:uni_roomie/screens/profile/EditProfilePage.dart';
 
-class EditProfilePage extends StatefulWidget {
+class ProfilePage extends StatefulWidget {
   @override
-  _EditProfilePageState createState() => _EditProfilePageState();
+  _ProfilePageState createState() => _ProfilePageState();
 }
 
-class _EditProfilePageState extends State<EditProfilePage> {
+class _ProfilePageState extends State<ProfilePage> {
   StreamSubscription<User> loginStateSubscription;
   String fullName;
   String email;
@@ -25,15 +28,11 @@ class _EditProfilePageState extends State<EditProfilePage> {
   DocumentReference course;
   String courseName;
   int yearOfStudy;
-  String profilePhoto;
   String userId;
-  TextEditingController nameController = new TextEditingController();
-  TextEditingController ageController = new TextEditingController();
-  TextEditingController yearOfStudyController = new TextEditingController();
-  Map<String, String> universities = Map<String, String>();
-  Map<String, String> courses = Map<String, String>();
-  List<QueryDocumentSnapshot> courseDocumentSnapshot;
-  List<QueryDocumentSnapshot> universityDocumentSnapshots;
+  String profilePhoto;
+
+  File _image;
+  final picker = ImagePicker();
 
   @override
   void initState() {
@@ -46,59 +45,29 @@ class _EditProfilePageState extends State<EditProfilePage> {
           ),
         );
       } else {
+        userId = fbUser.uid;
         DocumentReference documentReference =
             FirebaseFirestore.instance.collection('users').doc(fbUser.uid);
-        userId = fbUser.uid;
         documentReference.snapshots().listen((event) {
           setState(() {
             if (!mounted) return;
             fullName = event.data()["full_name"];
-            nameController.text = fullName;
             email = event.data()["email"];
             gender = event.data()["gender"];
             age = event.data()["age"];
-            ageController.text = age.toString();
             university = event.data()["university"];
             if (university != null)
               university.get().then((value) => setState(() {
-                    universityName = value.id;
+                    universityName = value.data()["name"];
                   }));
-            else
-              universityName == null;
             course = event.data()["course"];
             if (course != null)
               course.get().then((value) => setState(() {
-                    courseName = value.id;
+                    courseName = value.data()["name"];
                   }));
-            else
-              courseName = null;
             yearOfStudy = event.data()["yearOfStudy"];
-            yearOfStudyController.text = yearOfStudy.toString();
             profilePhoto = event.data()["profilePhoto"];
-          });
-        });
-        FirebaseFirestore.instance
-            .collection('courses')
-            .orderBy("name")
-            .get()
-            .then((results) {
-          courseDocumentSnapshot = results.docs;
-          results.docs.forEach((element) {
-            setState(() {
-              courses.putIfAbsent(element.id, () => element["name"]);
-            });
-          });
-        });
-        FirebaseFirestore.instance
-            .collection('universities')
-            .orderBy("name")
-            .get()
-            .then((results) {
-          universityDocumentSnapshots = results.docs;
-          results.docs.forEach((element) {
-            setState(() {
-              universities.putIfAbsent(element.id, () => element["name"]);
-            });
+            print(profilePhoto);
           });
         });
       }
@@ -106,24 +75,64 @@ class _EditProfilePageState extends State<EditProfilePage> {
     super.initState();
   }
 
+  Future getImage(ImageSource source) async {
+    final pickedFile = await picker.getImage(source: source);
+
+    setState(() {
+      if (pickedFile != null) {
+        _image = File(pickedFile.path);
+        UploadTask uploadTask = FirebaseStorage.instance
+            .ref('profilePhotos/$userId.png')
+            .putFile(_image);
+        uploadTask.then((snapshot) => {
+              snapshot.ref.getDownloadURL().then((value) => {
+                    FirebaseFirestore.instance
+                        .collection("users")
+                        .doc(userId)
+                        .update({"profilePhoto": value})
+                  })
+            });
+      } else {
+        print('No image selected.');
+      }
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
-    Map<String, String> genders = {
-      "Female": "Female",
-      "Male": "Male",
-      "Other": "Other"
-    };
     var authBloc = Provider.of<AuthBloc>(context, listen: false);
-    EditProfileList universityProfile = EditProfileList(
-        Icons.school, 'University', universities, universityName);
-    EditProfileList courseProfile =
-        EditProfileList(Icons.school, 'Course', courses, courseName);
-    EditProfileList genderProfile =
-        EditProfileList(Icons.person, 'Gender', genders, gender);
+    // set up the buttons
+    Widget cancelButton = FlatButton(
+      child: Text("Cancel"),
+      onPressed: () {
+        Navigator.of(context).pop();
+      },
+    );
+    Widget cameraButton = FlatButton(
+      child: Text("Camera"),
+      onPressed: () {
+        Navigator.of(context).pop();
+        getImage(ImageSource.camera);
+      },
+    );
+    Widget galleryButton = FlatButton(
+      child: Text("Gallery"),
+      onPressed: () {
+        Navigator.of(context).pop();
+        getImage(ImageSource.gallery);
+      },
+    );
+
+    // set up the AlertDialog
+    AlertDialog alert = AlertDialog(
+      title: Text("Photo Location"),
+      content: Text("Where do you want to select your new profile photo from?"),
+      actions: [cancelButton, galleryButton, cameraButton],
+    );
     return Scaffold(
       drawer: CustomDrawer(authBloc),
       appBar: AppBar(
-        title: Text('Edit Profile'),
+        title: Text('Your Profile'),
         centerTitle: true,
         backgroundColor: new Color.fromRGBO(69, 93, 122, 1),
       ),
@@ -148,6 +157,36 @@ class _EditProfilePageState extends State<EditProfilePage> {
                 ),
               ),
             ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Container(
+                  child: TextButton(
+                    onPressed: () => showDialog(
+                      context: context,
+                      builder: (BuildContext context) {
+                        return alert;
+                      },
+                    ),
+                    child: Row(
+                      children: [
+                        Text(
+                          'Change Avatar',
+                          style: TextStyle(
+                            color: Colors.black,
+                          ),
+                        ),
+                        SizedBox(width: 10),
+                        Icon(
+                          Icons.camera_alt,
+                          color: Colors.black,
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+              ],
+            ),
           ],
         ),
         Container(
@@ -166,29 +205,37 @@ class _EditProfilePageState extends State<EditProfilePage> {
           child: Column(
             children: [
               Container(
-                child: EditProfileTile(Icons.account_box, 'Name',
-                    nameController, TextInputType.text),
+                child: CustomProfileTile(Icons.account_box, 'Name',
+                    fullName == null ? "N/A" : fullName),
               ),
               Container(
-                child: genderProfile,
+                child: CustomProfileTile(Icons.alternate_email, 'Email',
+                    email == null ? "N/A" : email),
               ),
               Container(
-                child: EditProfileTile(
-                    Icons.grade, 'Age', ageController, TextInputType.number),
+                child: CustomProfileTile(
+                    Icons.person, 'Gender', gender == null ? "N/A" : gender),
+              ),
+              Container(
+                child: CustomProfileTile(
+                    Icons.grade, 'Age', age == null ? "N/A" : age.toString()),
               ),
               FittedBox(
                 child: Container(
-                  child: universityProfile,
+                  child: CustomProfileTile(Icons.school, 'University',
+                      universityName == null ? "N/A" : universityName),
                 ),
               ),
               FittedBox(
+                fit: BoxFit.contain,
                 child: Container(
-                  child: courseProfile,
+                  child: CustomProfileTile(Icons.bookmark, 'Course',
+                      courseName == null ? "N/A" : courseName),
                 ),
               ),
               Container(
-                child: EditProfileTile(Icons.trending_up, 'Year Of Study',
-                    yearOfStudyController, TextInputType.number),
+                child: CustomProfileTile(Icons.trending_up, 'Year Of Study',
+                    yearOfStudy == null ? "N/A" : yearOfStudy.toString()),
               ),
               Container(
                 padding: const EdgeInsets.fromLTRB(16.0, 16.0, 16.0, 16.0),
@@ -196,26 +243,10 @@ class _EditProfilePageState extends State<EditProfilePage> {
                   shape: RoundedRectangleBorder(
                       borderRadius: BorderRadius.circular(18.0)),
                   onPressed: () {
-                    FirebaseFirestore.instance
-                        .collection("users")
-                        .doc(userId)
-                        .update({
-                      "age": int.parse(ageController.text),
-                      "course": courseDocumentSnapshot
-                          .firstWhere((element) =>
-                              element.id == courseProfile.selectedValue)
-                          .reference,
-                      "full_name": nameController.text,
-                      "gender": genderProfile.selectedValue,
-                      "university": universityDocumentSnapshots
-                          .firstWhere((element) =>
-                              element.id == universityProfile.selectedValue)
-                          .reference,
-                      "yearOfStudy": int.parse(yearOfStudyController.text)
-                    });
                     Navigator.push(
                       context,
-                      MaterialPageRoute(builder: (context) => ProfilePage()),
+                      MaterialPageRoute(
+                          builder: (context) => EditProfilePage()),
                     );
                   },
                   color: new Color.fromRGBO(249, 89, 89, 1),
@@ -225,12 +256,12 @@ class _EditProfilePageState extends State<EditProfilePage> {
                       mainAxisAlignment: MainAxisAlignment.center,
                       children: [
                         Text(
-                          'Save Changes',
+                          'Edit Profile',
                           style: TextStyle(fontSize: 20.0),
                         ),
                         SizedBox(width: 10),
                         Icon(
-                          Icons.save_alt,
+                          Icons.edit,
                           color: Colors.black,
                         ),
                       ],
