+++ lib/services/AuthService.dart
@@ -4,7 +4,9 @@ class AuthService {
   final _auth = FirebaseAuth.instance;
 
   Future<UserCredential> signInWithCredential(AuthCredential credential) =>
-  _auth.signInWithCredential(credential);
+      _auth.signInWithCredential(credential);
+
   Future<void> logout() => _auth.signOut();
+
   Stream<User> get currentUser => _auth.authStateChanges();
 }
